requirements.txt
beautifulsoup4
requests
selenium
pillow
tkinter
html5lib
lxml
cryptography